Delete target directory before generation
Also only create api directory if needed.
