--- src/test/java/mada/fixture/DirectoryDeleter.java
@@ -1,48 +1,0 @@
-package mada.fixture;
-
-import java.io.IOException;
-import java.nio.file.FileVisitResult;
-import java.nio.file.FileVisitor;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.attribute.BasicFileAttributes;
-
-/**
- * Deletes a directory.
- */
-public final class DirectoryDeleter {
-  private DirectoryDeleter() {
-  }
-
-  /**
-   * Deletes directory.
-   *
-   * @param dir directory to delete
-   * @throws IOException if the operation fails
-   */
-  public static void delete(Path dir) throws IOException {
-    Files.walkFileTree(dir, new FileVisitor<Path>() {
-      @Override
-      public FileVisitResult preVisitDirectory(Path dir, BasicFileAttributes attrs) throws IOException {
-        return FileVisitResult.CONTINUE;
-      }
-
-      @Override
-      public FileVisitResult visitFile(Path file, BasicFileAttributes attrs) throws IOException {
-        Files.delete(file);
-        return FileVisitResult.CONTINUE;
-      }
-
-      @Override
-      public FileVisitResult visitFileFailed(Path file, IOException exc) throws IOException {
-        return FileVisitResult.CONTINUE;
-      }
-
-      @Override
-      public FileVisitResult postVisitDirectory(Path dir, IOException exc) throws IOException {
-        Files.delete(dir);
-        return FileVisitResult.CONTINUE;
-      }
-    });
-  }
-}
