+++ src/test/java/mada/fixture/TestIterator.java
@@ -15,6 +15,7 @@ import org.junit.platform.suite.api.Suite;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import dk.mada.jaxrs.utils.DirectoryDeleter;
 import dk.mada.logging.LoggerConfig;
 
 @Suite
